+++ index.d.ts
@@ -18,6 +18,7 @@ declare let __webpack_require__: (moduleId: string) => any;
 
 declare const MODULE: string;
 declare const CSP_NONCE_STORE: string;
+declare const LANG_PACKS: string;
 
 declare var PATH: Dictionary<CanUndef<string>>;
 declare var PUBLIC_PATH: CanUndef<string>;
@@ -27,6 +28,8 @@ declare const TPLS: Dictionary<Dictionary<Function>>;
 
 declare const DS: CanUndef<DesignSystem>;
 declare const BLOCK_NAMES: CanUndef<string[]>;
+
+declare const BUILD_MODE: CanUndef<string>;
 
 declare const DS_COMPONENTS_MODS: CanUndef<{
 	[name: string]: Nullable<Array<string | boolean | number>>;