+++ src/lang/index.ts
@@ -8,35 +8,22 @@
 
 import Super from '@v4fire/core/lang';
 
+import keysetsCollector from 'lang/engines';
+
 export * from '@v4fire/core/lang';
 
 const langPacs = {
 	...Super
 };
 
-// @context: ['src', '@super']
-
-const
-	// @ts-ignore (require)
-	ctx = require.context('src', true, /\.i18n\/.*\.js$/);
-
-ctx.keys().forEach((path) => {
-	const
-		parsedPath = /\/[^/]*?\.i18n\/(.*?)\.js$/i.exec(path);
-
-	if (parsedPath != null) {
-		const [_, lang] = parsedPath;
-		langPacs[lang] ??= {};
-
-		Object.keys(ctx(path)).forEach((keysetName) => {
-			langPacs[lang][keysetName] = {
-				...langPacs[lang][keysetName],
-				...ctx(path)[keysetName]
-			};
-		});
-	}
+Object.entries(keysetsCollector()).forEach(([lang, keysets]) => {
+	Object.entries(keysets).forEach(([keysetName, keyset]) => {
+		langPacs[lang] = langPacs[lang] ?? {};
+		langPacs[lang][keysetName] = {
+			...langPacs[lang][keysetName],
+			...keyset
+		};
+	});
 });
-
-// @endcontext
 
 export default langPacs;