+++ src/super/i-static-page/modules/ss-helpers/page.spec.js
@@ -50,6 +50,8 @@ Object.defineProperty(window, 'GLOBAL_NONCE', {
 \tvalue: undefined
 });
 
+var LANG_PACKS = {};
+
 var PATH = Object.create(null);
 var PUBLIC_PATH = undefined;